api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
PyPDF2
python-multipart